+++ src/allpress/settings.py
@@ -1,5 +1,6 @@
 
-CLASSIFICATION_MODELS_PATH = "C:\\Users\\preit\\OneDrive\\Desktop\\coding projects\\Allpress2\\src\\allpress\\models"
+CLASSIFICATION_MODELS_PATH = "C:\\Users\\preit\\OneDrive\\Desktop\\coding projects\\Allpress2\\src\\allpress\\models\\autoencoders"
+FAISS_INDEX_PATH = "C:\\Users\\preit\\OneDrive\\Desktop\\coding projects\\Allpress2\\src\\allpress\\models\\faiss"
 
 DATABASE_USERNAME = "test"
 DATABASE_PASSWORD = "default"